Json Process

+++ Assets/Scripts/GlobalManage.cs
@@ -9,14 +9,15 @@ namespace Global
         public static GlobalManage Instance { get; private set;}
 
         [SerializeField]
-        private float m_speed;
+        private int m_speed;
         [SerializeField]
         private bool m_timer_flag = true;
         [SerializeField]
         private float[] m_distances;
 
 
-        public float speed { get { return m_speed; } set { m_speed = value; } }
+
+        public int speed { get { return m_speed; } set { m_speed = value; } }
         public bool timer_flag { get { return m_timer_flag; } set { m_timer_flag = value; } }
         public float [] distance_Enem { get { return m_distances; }set { m_distances = value; } }
